Move Redux connect from App to components
